% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_model.R
\name{build_cross_validation_time_stitch_classification_models}
\alias{build_cross_validation_time_stitch_classification_models}
\title{build an xgboost cross validation classification model with k-fold cross-validation for each featureset provided, assumed that the classification is defined by the previoud model}
\usage{
build_cross_validation_time_stitch_classification_models(
  base_name,
  base_predictor,
  target_list,
  feature_list,
  q_thresh = 0.05
)
}
\arguments{
\item{base_predictor}{\itemize{
\item classification model used the emprical classifier at target age
}}

\item{target_list}{\itemize{
\item list of patients, each represents the next (descending) age group.
}}

\item{feature_list}{\itemize{
\item list of features, each represents the next (descending) age group.
}}

\item{q_thresh}{\itemize{
\item score quantile threshold for target classification of 1
}}

\item{base_age}{\itemize{
\item age of initial predictor
}}
}
\value{
a list of predictors, according to provided target_list
}
\description{
build an xgboost cross validation classification model with k-fold cross-validation for each featureset provided, assumed that the classification is defined by the previoud model
}
