% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_model.R
\name{build_mortality_markov_model}
\alias{build_mortality_markov_model}
\title{build an Markov probability model from multi-age prediction models
To implement this, all patients at a given age will be binned according to their model score (using quantiles).
Each bin is assigned a state, and we are computing the probability for traversing from each state to the next model state'}
\usage{
build_mortality_markov_model(
  models,
  follow_time,
  outcome,
  step,
  qbins = seq(0, 1, by = 0.05)
)
}
\arguments{
\item{models}{\itemize{
\item list of prediction models (output of build_cross_validation_time_stitch_classification_models)
}}

\item{follow_time}{\itemize{
\item list of data.frames per age, defining for each patient how much follow-up time they had. Dataframe must include the following columns:
\item id
\item time_in_system
\item target_class - the classification if known for outcome years or within step (i.e. patient died). NA for unknown step outcome.
}}

\item{outcome}{\itemize{
\item time from oldest model (first) to target outcome
}}

\item{step}{\itemize{
\item time between prediction models
}}

\item{qbins}{\itemize{
\item quantile bin size of prediction score for which the markov model will define a state
}}
}
\value{
a list of markov models (per age), each is a list with the following members:
\itemize{
\item model - data.frame containing the probability for each quantile(score) bin to reach each of the target_classes provided in the oldest model.
\item local.model - data.frame containing the probability for each quantile(score) bin to reach each of the quantile(score) bins of the next model by age.
\item qbins -  bins
\item target - data frame containing the target bin for this age model (to be used as outcome for the younger age model)
}
}
\description{
build an Markov probability model from multi-age prediction models
To implement this, all patients at a given age will be binned according to their model score (using quantiles).
Each bin is assigned a state, and we are computing the probability for traversing from each state to the next model state'
}
