% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease_markov_model.R
\name{.disease_markov_model_for_stitch_model}
\alias{.disease_markov_model_for_stitch_model}
\title{build an Markov probability model from multi-age prediction models}
\usage{
.disease_markov_model_for_stitch_model(
  markov,
  model,
  step,
  qbins,
  required_conditions,
  min_obs_for_estimate = 10
)
}
\arguments{
\item{markov}{\itemize{
\item the markov model computed for the next age (older). the states in this age will be mapped to the states in this markov layer
}}

\item{model}{\itemize{
\item prediction model (output of build_cross_validation_time_stitch_classification_models)
}}

\item{step}{\itemize{
\item time between prediction models
}}

\item{qbins}{\itemize{
\item quantile bin size of prediction score for which the markov model will define a state
}}

\item{follow}{\itemize{
\item data.frames  defining for each patient how much follow-up time they had. Dataframe must include the following columns:
\item id
\item time_in_system
\item target_class - the classification if known within step (i.e. patient died). NA for unknown step outcome.
}}
}
\value{
a markov model, a list with the following members:
\itemize{
\item model - matrix containing the probability for each quantile(score) bin to reach each of the target_classes provided in the oldest model.
\item local.model - data.frame containing the probability for each quantile(score) bin to reach each of the quantile(score) bins of the next model by age.
\item qbins -  bins
\item target - data frame containing the target bin for this age model (to be used as outcome for the younger age model)
}
}
\description{
build an Markov probability model from multi-age prediction models
}
