% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease_markov_model.R
\name{.mldpEHR.disease_empirical_prob_for_disease}
\alias{.mldpEHR.disease_empirical_prob_for_disease}
\title{this method computes the probability for future disease according to the availablitiy of data
by basically computing a stitched markov model
as score availability is not unbiased (patients with less measurements tend to be healthier),
the probaility will be computed seperately for patients with / without score
we will define the following modes:
\itemize{
\item 0 has required conditions at young age and known score / outcome at older age
\item 1 has required conditions at young age and no restrictions at older age
\item 2 no restrictions at young/older age
}}
\usage{
.mldpEHR.disease_empirical_prob_for_disease(
  population,
  step,
  required_conditions = "id==id"
)
}
\arguments{
\item{population}{\itemize{
\item list of data.frames of all the patients in the system going back in time. For example the first data.frame represents age 80, next is 75 and so forth.
Each patient data.frame contains the following columns:
\item patient id
\item sex
\item target_class (does this patient have a known outcome)
\item disease age of disease
\item death age of death
\item followup
\item any other columns that can be used in required conditions
}}

\item{step}{\itemize{
\item time between populations
}}

\item{required_conditions}{\itemize{
\item conditions that will be applied on patients that will be have a prediction score
}}
}
\description{
this method computes the probability for future disease according to the availablitiy of data
by basically computing a stitched markov model
as score availability is not unbiased (patients with less measurements tend to be healthier),
the probaility will be computed seperately for patients with / without score
we will define the following modes:
\itemize{
\item 0 has required conditions at young age and known score / outcome at older age
\item 1 has required conditions at young age and no restrictions at older age
\item 2 no restrictions at young/older age
}
}
