% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease_markov_model.R
\name{.mldpEHR.disease_expected}
\alias{.mldpEHR.disease_expected}
\title{calculate expected number of disease patients}
\usage{
.mldpEHR.disease_expected(index, population_count, edp)
}
\arguments{
\item{population_count}{\itemize{
\item data.frame containing age, sex, and the number of patients available
}}

\item{empirical_disease_prob}{\itemize{
\item output of .mldpEHR.disease_empirical_prob_for_disease
}}

\item{step}{\itemize{
\item time between populations
}}
}
\description{
calculate expected number of disease patients
}
