% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_model.R
\name{.mortality_markov_model_for_stitch_model}
\alias{.mortality_markov_model_for_stitch_model}
\title{build an Markov probability model from multi-age prediction models}
\usage{
.mortality_markov_model_for_stitch_model(markov, model, follow, step, qbins)
}
\arguments{
\item{markov}{\itemize{
\item the markov model computed for the next age (older). the states in this age will be mapped to the states in this markov layer
}}

\item{model}{\itemize{
\item prediction model (output of build_cross_validation_time_stitch_classification_models)
}}

\item{follow}{\itemize{
\item data.frames  defining for each patient how much follow-up time they had. Dataframe must include the following columns:
\item id
\item time_in_system
\item target_class - the classification if known within step (i.e. patient died). NA for unknown step outcome.
}}

\item{step}{\itemize{
\item time between prediction models
}}

\item{qbins}{\itemize{
\item quantile bin size of prediction score for which the markov model will define a state
}}
}
\value{
\subsection{a markov models, a list with the following members:}{
}
}
\description{
build an Markov probability model from multi-age prediction models
}
