% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{load_disease_example_data}
\alias{load_disease_example_data}
\title{Load a mock dataset for disease prediction}
\usage{
load_disease_example_data(N = 1000, num_age_groups = 5)
}
\arguments{
\item{N}{The number of patients to generate}

\item{num_age_groups}{The number of age groups to generate}
}
\value{
an MldpEHR object
}
\description{
Load a mock dataset for disease prediction
}
\examples{
disease_data <- load_disease_example_data(1000)

}
