% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{load_mortality_example_data}
\alias{load_mortality_example_data}
\title{Load a mock dataset for mortality prediction}
\usage{
load_mortality_example_data(N = 1000, num_age_groups = 5)
}
\arguments{
\item{N}{The number of patients to generate}

\item{num_age_groups}{The number of age groups to generate}
}
\value{
an MldpEHR object
}
\description{
Load a mock dataset for mortality prediction
}
\examples{
mortality_data <- load_mortality_example_data(1000)

}
