% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease_markov_model.R
\name{mldp_disease_expected}
\alias{mldp_disease_expected}
\title{calculate expected number of disease patients}
\usage{
mldp_disease_expected(index, population_count, edp)
}
\arguments{
\item{index}{\itemize{
\item index of entry in the empirical disease prob of the current age to start propagation from
}}

\item{population_count}{\itemize{
\item data.frame containing age, sex, and the number of patients available
}}

\item{edp}{\itemize{
\item empirical disease prob, output of mldp_disease_empirical_prob_for_disease
}}
}
\description{
calculate expected number of disease patients
}
