% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease_markov_model.R
\name{mldp_disease_markov}
\alias{mldp_disease_markov}
\title{Build a Markov probability model from multi-age prediction models of a disease}
\usage{
mldp_disease_markov(
  models,
  target_years,
  qbins = seq(0, 1, by = 0.05),
  required_conditions = NULL
)
}
\arguments{
\item{models}{a list of prediction models (output of \code{mldp_disease_multi_age_predictors}).}

\item{target_years}{the number of years to predict the disease for (e.g. 5 years).}

\item{qbins}{a vector of quantile bins for the prediction score from which the markov model will define a state. Default is to use 10 equally sized quantiles.}

\item{required_conditions}{a string with an expression for any filter to apply to the patients to filter out from model computation. Can be used to filter out patients with missing data or limiting the time window. See \code{\link{dplyr::filter}} for more details.}
}
\value{
a list of with the following members:
\itemize{
\item{prob: }{a data frame containing the life-long probability to get the disease ("disease"), die ("death") die with the disease ("death_disease"), or not get the disease and live ("healthy") for each age, gender and score bin (quantile of score).}#'
\item{models: }{a list of matrices containing the probability for each quantile(score) bin to reach
each of the quantile(score) bins of the next model by age.}
}
}
\description{
All patients at a given age will be binned according to their model score (using quantiles).
Each bin is assigned a state, and we are computing the probability for traversing from
each state to the next model state. \cr
Patients with missing score are also included for this model to reflect actual population numbers.
}
\examples{


# Load a small example data
disease_data <- load_disease_example_data(N = 100, num_age_groups = 3)

# Build predictors
predictors <- mldp_disease_multi_age_predictors(disease_data@patients, disease_data@features, target_years = 5, nfolds = 2)

markov <- mldp_disease_markov(predictors, 5, qbins = seq(0, 1, by = 0.1))

# plot the life-long probability for the disease for each age and gender

library(ggplot2)
library(dplyr)

markov$prob \%>\%
    mutate(sex = factor(c("male", "female")[sex])) \%>\%
    ggplot(aes(x = sbin, y = disease + disease_death, colour = sex, group = sex)) +
    geom_point() +
    geom_line() +
    ylab("p(disease)") +
    xlab("score bin") +
    scale_color_manual(name = "", values = c("male" = "blue", "female" = "red")) +
    facet_wrap(~age, nrow = 1) +
    theme_bw() +
    theme(axis.text.x = element_text(angle = 90, hjust = 1))

}
