% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mldp_download_example_data}
\alias{mldp_download_example_data}
\title{Download example data}
\usage{
mldp_download_example_data(
  output_dir = file.path(getwd(), "examples"),
  datasets = c("longevity", "diabetes"),
  overwrite = FALSE,
  timeout = 60 * 60 * 2
)
}
\arguments{
\item{output_dir}{Directory to save the data to. Defaults to "examples" in the current working directory.}

\item{datasets}{A character vector of datasets to download. Defaults to "longevity" and "diabetes".}

\item{overwrite}{Overwrite existing files. Defaults to FALSE.}

\item{timeout}{Timeout in seconds for downloading files. Defaults to 2 hours. If you are having trouble downloading the files, try increasing this value.}
}
\value{
Nothing
}
\description{
Download example data. See description of the datasets below.
}
\section{longevity.patients}{

A dataset for patient survival.

This dataset is a list of 9 data frames, representing the population of patients for
each age starting at 80 and going down to 40 at 5 year interavls. For each patients,
the age at death is provided.
\cr

Each dataframe has the following 5 columns:
\itemize{
\item{id: }{unique patient id}
\item{age: }{age of patient}
\item{sex: }{1 for male, 2 for female}
\item{death: }{age at death, NA if patient doesn't die by end of followup}
\item{has_cbc: }{binary indicating if patient has available data}
\item{followup: }{available followup time in years}
}
\cr
}

\section{longevity.features}{

A dataset for patient survival features.

This dataset is a list of 9 data frames, representing the features of patients available in longevity.patients.
Similar to longevity.patients, each data frame is for a given age population,
starting at 80 and going down to 40 at 5 year interavls
\cr
Each data frame has 86 columns. The features include sex and common labs, quantile normalized values.
}

\section{diabetes.patients}{

A dataset for patient disease outcome.

This dataset is a list of 9 data frames, representing the population of patients for
each age starting at 80 and going down to 40 at 5 year interavls. For each patients,
the age at disease is provided.
\cr
Each dataframe with the following 5 columns:
\itemize{
\item{id: }{unique patient id}
\item{age: }{age of patient}
\item{sex: }{1 for male, 2 for female}
\item{death: }{age at death, NA if patient doesn't die by end of follow-up}
\item{disease: }{age at diabetes disease, NA if patient doesn't get sick by end of follow-up}
\item{has_cbc: }{binary indicating if patient has available data}
\item{followup: }{available follow-up time in years}
}
\cr
}

\section{diabetes.features}{

A dataset for patient diabetes features.

This dataset is a list of 9 data frames, representing the features of patients available in diabetes.patients.
Similar to diabetes.patients, each data frame is for a given age population,
starting at 80 and going down to 40 at 5 year interavls.
\cr
Each data frame has 86 columns. The features include sex and common labs, quantile normalized values.
}

\examples{
\donttest{
mldp_download_example_data()
}

\dontshow{
unlink("examples", recursive = TRUE)
}

}
