% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mldp_load_data}
\alias{mldp_load_data}
\title{Load example data}
\usage{
mldp_load_data(dataset, data_dir = file.path(getwd(), "examples"))
}
\arguments{
\item{dataset}{A character vector of the dataset to load, e.g. "longevity" or "diabetes".}

\item{data_dir}{Directory to load the data from. Defaults to "examples" in the current working directory.
You can download the data using \code{mldpEHR::download_example_data()}.}
}
\value{
an MldpEHR object
}
\description{
Load example data that was downloaded with \code{mldpEHR::download_example_data()}. See description of the datasets below.
}
\section{longevity.patients}{

A dataset for patient survival.

This dataset is a list of 9 data frames, representing the population of patients for
each age starting at 80 and going down to 40 at 5 year interavls. For each patient,
the age at death is provided.
\cr

Each data frame has the following 5 columns:
\itemize{
\item{id: }{unique patient id}
\item{age: }{age of patient}
\item{sex: }{1 for male, 2 for female}
\item{death: }{age at death, NA if patient doesn't die by end of followup}
\item{has_cbc: }{binary indicating if patient has required feature data}
\item{followup: }{available followup time in years}
}
\cr
}

\section{longevity.features}{

A dataset for patient survival features.

This dataset is a list of 9 data frames, representing the features of patients available in longevity.patients.
Similar to longevity.patients, each data frame is for a given age population,
starting at 80 and going down to 40 at 5 year interavls
\cr
Each data frame has 86 columns. The features include sex and common labs, quantile normalized values.
}

\section{diabetes.patients}{

A dataset for patient disease outcome.

This dataset is a list of 9 data frames, representing the population of patients for
each age starting at 80 and going down to 40 at 5 year interavls. For each patient,
the age at disease is provided.
\cr
Each dataframe with the following 5 columns:
\itemize{
\item{id: }{unique patient id}
\item{age: }{age of patient}
\item{sex: }{1 for male, 2 for female}
\item{death: }{age at death, NA if patient doesn't die by end of follow-up}
\item{disease: }{age at diabetes disease, NA if patient doesn't get sick by end of follow-up}
\item{has_cbc: }{binary indicating if patient has available data}
\item{followup: }{available follow-up time in years}
}
\cr
}

\section{diabetes.features}{

A dataset for patient diabetes features.

This dataset is a list of 9 data frames, representing the features of patients available in diabetes.patients.
Similar to diabetes.patients, each data frame is for a given age population,
starting at 80 and going down to 40 at 5 year interavls.
\cr
Each data frame has 86 columns. The features include sex and common labs, quantile normalized values.
}

\examples{
\dontrun{
mldp_download_example_data(file.path(getwd(), "examples"))
longevity_data <- mldp_load_data("longevity", file.path(getwd(), "examples"))
diabetes_data <- mldp_load_data("diabetes", file.path(getwd(), "examples"))
}

\dontshow{
unlink("examples", recursive = TRUE)
}

}
