% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_model.R
\name{mldp_model_export}
\alias{mldp_model_export}
\title{Export predictors, stripping all patient data, for future use in inference of external data.}
\usage{
mldp_model_export(models)
}
\arguments{
\item{models}{a list of prediction models (output of \code{mldp_mortality_multi_age_predictors} or \code{mldp_disease_multi_age_predictors}).}
}
\value{
a list of with the following members:
\itemize{
\item{model: }{a list of xgboost models, for each fold}
\item{feature_names: }{a vector of expected features used in model}
\item{age: }{the age of patients used in the model}
}
}
\description{
Export predictors, stripping all patient data, for future use in inference of external data.
}
\examples{

library(ggplot2)
library(dplyr)

mortality <- load_mortality_example_data(N = 100, num_age_groups = 3)
predictors <- mldp_mortality_multi_age_predictors(
    mortality@patients,
    mortality@features,
    survival_years = 5,
    nfolds = 2,
    q_thresh = 0.2,
    nthread = 2 # CRAN allows only 2 cores
)
predictors <- mldep_model_export(predictors)

}
