% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_markov_model.R
\name{mldp_mortality_markov}
\alias{mldp_mortality_markov}
\title{Build a Markov probability model from multi-age prediction models of mortality}
\usage{
mldp_mortality_markov(
  models,
  survival_years,
  qbins = seq(0, 1, by = 0.05),
  required_conditions = NULL
)
}
\arguments{
\item{models}{a list of prediction models (output of \code{mldp_mortality_multi_age_predictors}).}

\item{survival_years}{The number of survival years the model aimed to predict for the oldest age group (see \code{mldp_mortality_multi_age_predictors}).}

\item{qbins}{a vector of quantile bins for the prediction score from which the markov model will define a state. Default is to use 10 equally sized quantiles.}

\item{required_conditions}{a string with an expression for any filter to apply to the patients to filter out from model computation. Can be used to filter out patients with missing data or limiting the time window. See \code{\link[dplyr]{filter}} for more details.}
}
\value{
a list of with the following members:
\itemize{
\item{prob: }{a data frame containing the probability to survive ("alive") or die ("death") before the oldest age group + survival_years for each age, gender and score bin (quantile of score).}
\item{models: }{a list of matrices containing the probability for each quantile(score) bin to reach
each of the quantile(score) bins of the next model by age.}
}
}
\description{
All patients at a given age will be binned according to their model score (using quantiles).
Each bin is assigned a state, and we are computing the probability for traversing from
each state to the next model state. \cr
Patients with missing score are also included for this model to reflect actual population numbers.
}
\examples{

library(ggplot2)
library(dplyr)

mortality <- load_mortality_example_data(N = 100, num_age_groups = 3)
predictors <- mldp_mortality_multi_age_predictors(
    mortality@patients,
    mortality@features,
    survival_years = 5,
    nfolds = 2,
    q_thresh = 0.2,
    nthread = 2 # CRAN allows only 2 cores
)

markov <- mldp_mortality_markov(predictors, 5, qbins = seq(0, 1, by = 0.1))

head(markov$prob)

# plot the probability of survival for each age group
markov$prob \%>\%
    mutate(sex = factor(c("male", "female")[sex])) \%>\%
    ggplot(aes(x = sbin, y = alive, colour = sex, group = sex)) +
    geom_point() +
    geom_line() +
    ylab("p(survival)") +
    xlab("score bin") +
    scale_color_manual(name = "", values = c("male" = "blue", "female" = "red")) +
    facet_wrap(~age, nrow = 1) +
    theme_bw() +
    theme(axis.text.x = element_text(angle = 90, hjust = 1))

}
