% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-prediction_model.R
\name{mldp_plot_multi_age_predictors_ecdf}
\alias{mldp_plot_multi_age_predictors_ecdf}
\title{Plot an ecdf of the score of each model for colored by outcome}
\usage{
mldp_plot_multi_age_predictors_ecdf(predictors)
}
\arguments{
\item{predictors}{a list of predictors, output of \code{mldp_mortality_multi_age_predictors}}
}
\value{
a ggplot object
}
\description{
For the oldest age the outcome is death, for younger ages the outcome is the score quantile of the next age group.
}
\examples{
# Load the example data
data <- load_mortality_example_data()
predictors <- mldp_mortality_multi_age_predictors(mortality@patients, mortality@features, survival_years = 5, nfolds = 3, q_thresh = 0.2)
mldp_plot_multi_age_predictors_ecdf(predictors)

}
