% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-prediction_model.R
\name{mldp_plot_multi_age_predictors_shap}
\alias{mldp_plot_multi_age_predictors_shap}
\title{Plot SHAP (SHapley Additive exPlanations) values for a model}
\usage{
mldp_plot_multi_age_predictors_shap(predictor_shap)
}
\arguments{
\item{predictor_shap}{output of \code{mldp_prediction_model_features}}

\item{n_features}{the number of features to plot}

\item{n_patients}{the number of patients to plot}
}
\value{
a ggplot object
}
\description{
Plot SHAP (SHapley Additive exPlanations) values for a model
}
\examples{
# Load the example data
data <- load_mortality_example_data()
predictors <- mldp_mortality_multi_age_predictors(mortality@patients, mortality@features, survival_years = 5, nfolds = 3, q_thresh = 0.2)
predictor_shap <- mldp_prediction_model_features(predictors$`80`)
mldp_plot_prediction_model_features(predictor_shap)

}
