% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-prediction_model.R
\name{mldp_plot_shap}
\alias{mldp_plot_shap}
\title{Plot SHAP (SHapley Additive exPlanations) values for a model}
\usage{
mldp_plot_shap(
  predictor_shap,
  n_patients = 1000,
  n_features = NULL,
  point_size = 0.01,
  point_alpha = 0.3,
  ...
)
}
\arguments{
\item{predictor_shap}{output of \code{mldp_prediction_model_features}}

\item{n_patients}{the number of patients to plot. Default is to sample 1000 patients. If \code{n_patients} is greater than the number of patients in the data, all patients are plotted.}

\item{n_features}{the number of features to plot. Features will be sorted by the average absolute SHAP. If \code{NULL} all features are plotted.}

\item{point_size}{the size of the points in the plot. Default is 0.01.}

\item{point_alpha}{the transparency of the points in the plot. Default is 0.3.}

\item{...}{
  Arguments passed on to \code{\link[ggplot2:geom_point]{ggplot2::geom_point}}
  \describe{
    \item{\code{mapping}}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}
    \item{\code{data}}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}
    \item{\code{stat}}{The statistical transformation to use on the data for this
layer, either as a \code{ggproto} \code{Geom} subclass or as a string naming the
stat stripped of the \code{stat_} prefix (e.g. \code{"count"} rather than
\code{"stat_count"})}
    \item{\code{position}}{Position adjustment, either as a string naming the adjustment
(e.g. \code{"jitter"} to use \code{position_jitter}), or the result of a call to a
position adjustment function. Use the latter if you need to change the
settings of the adjustment.}
    \item{\code{na.rm}}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}
    \item{\code{show.legend}}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}
    \item{\code{inherit.aes}}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}
  }}
}
\value{
a ggplot object
}
\description{
Plot SHAP (SHapley Additive exPlanations) values for a model
}
\examples{

mortality <- load_mortality_example_data(N = 100, num_age_groups = 3)
predictors <- mldp_mortality_multi_age_predictors(
    mortality@patients,
    mortality@features,
    survival_years = 5,
    nfolds = 2,
    q_thresh = 0.2,
    nthread = 2 # CRAN allows only 2 cores
)

predictor_shap <- mldp_model_features(predictors$`80`)
mldp_plot_shap(predictor_shap)

}
