% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MldpEHR.R
\docType{class}
\name{MldpEHR-class}
\alias{MldpEHR-class}
\title{Data for mldpEHR models}
\description{
Data for mldpEHR models
}
\section{Slots}{

\describe{
\item{\code{patients}}{list of data.frames of all the patients in the system going back in time. For example the first data.frame represents age 80, next is 75 and so forth. See the creation function for more details.}

\item{\code{features}}{a list of data.frames of all the features for the patients in the system going back in time. For example the first data.frame represents age 80, next is 75 and so forth. See the creation function for more details.}

\item{\code{age_groups}}{(optional) labels for the age groups}
}}

