% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mldpEHR-package.R
\docType{package}
\name{mldpEHR-package}
\alias{mldpEHR}
\alias{mldpEHR-package}
\title{mldpEHR: Markovian Lifelong Mortality and Disease Predisposition models for electronic health records.}
\description{
A toolkit for building and testing lifelong prediction models for all cause mortality (longevity) and disease.
}
\author{
\strong{Maintainer}: Netta Mendelson-Cohen \email{netta.mendelsoncohen@weizmann.ac.il}

Authors:
\itemize{
  \item Aviezer Lifshitz \email{aviezer.lifshitz@weizmann.ac.il}
}

Other contributors:
\itemize{
  \item Weizmann Institute of Science [copyright holder]
}

}
\keyword{internal}
