% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_model.R
\name{mldpEHR.cv_model_features}
\alias{mldpEHR.cv_model_features}
\title{analyze feature significance for a k-fold cross validation model using shaply values}
\usage{
mldpEHR.cv_model_features(predictor)
}
\arguments{
\item{predictor}{\itemize{
\item classification model trained by mldpEHR.cv_train_outcome
}}

\item{nrounds}{\itemize{
\item number of training rounds
}}
}
\value{
a data.frame containing all features and their statistics:
\itemize{
\item model - list of xgboost models, for each fold
\item train - data.frame containing the patients id, fold, target class and predicted value in training (each id was used in nfolds-1 for training)
\item test - data.frame containing the patients id, fold, target class and predicted value in testing (each id was tested once in the fold it was not used for training)
\item xgboost_params - the set of parameters used in xgboost
\item nrounds - number of training iterations conducted
}
}
\description{
analyze feature significance for a k-fold cross validation model using shaply values
}
\examples{
library(dplyr)
library(ggplot2)
target <- data.frame(id = 1:1000, sex = rep(c("male", "female"), 500), target_class = rep(c(0, 1), each = 500))
features <- data.frame(id = 1:500, a = rnorm(500), b = rnorm(500)) \%>\%
    bind_rows(
        data.frame(id = 501:1000, a = rnorm(500, mean = 0.5, sd = 2), b = rnorm(500, mean = -0.5, sd = 2))
    )
predictor <- mldpEHR.cv_train_outcome(target, features, folds = 3)
}
