% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_model.R
\name{mldpEHR.prediction_model_features}
\alias{mldpEHR.prediction_model_features}
\title{analyze feature significance for a k-fold cross validation model using shaply values}
\usage{
mldpEHR.prediction_model_features(predictor)
}
\arguments{
\item{predictor}{\itemize{
\item classification model trained by mldpEHR.cv_train_outcome
}}
}
\value{
a list with the 3 objects
\itemize{
\item summary - data frame containing for each feature the mean absolute shaply value for the feature across all training data
\item shap_by_patient - data frame containing for each patient and feature the feature value and mean shap value across all training folds
\item shap_by_fold - similar to shap_by_patient, but for each fold seperately
}
}
\description{
analyze feature significance for a k-fold cross validation model using shaply values
}
\examples{
N <- 1000
patients <- list(data.frame(
    id = 1:N,
    sex = rep(c(1, 2), N / 2),
    age = 80 - .x * 5,
    death = c(rep(NA, 0.2 * N), rep(82, 0.8 * N)),
    followup = .x * 5 + 5
))
features <- list(data.frame(
    id = 1:N,
    a = c(rnorm(0.2 * N), rnorm(0.8 * N, mean = 2, sd = 1)),
    b = rep(c(rnorm(N / 4), rnorm(N / 4, mean = 3)), 2)
))
predictor <- mldpEHR.mortality_multi_age_predictors(patients, features, 5, 3, q_thresh = 0.05)
predictor_features <- mldpEHR.prediction_model_features(predictor[[1]])
}
