% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MldpEHR.R
\name{MldpEHR}
\alias{MldpEHR}
\title{Create MldpEHR object}
\usage{
MldpEHR(patients, features, age_groups = NULL)
}
\arguments{
\item{patients}{list of data.frames of all the patients in the system going back in time. For example the first data.frame represents age 80, next is 75 and so forth. Each patient data.frame contains the following columns:
\itemize{
\item{id: }{unique patient id}
\item{age: }{age of patient. All patients in the data.frame must be the same age}
\item{sex: }{1 for male, 2 for female}
\item{death: }{age at death, NA if patient doesn't die by end of followup}
\item{followup: }{available followup time (in years) for this patient - time until end of database or until patient exists the system (not due to death)}
}\cr

The data frame can contain any additional columns required for patient filtering in the future.\cr

Note that every age group except the last one must have at least two patients per sex which exist in the next age group. This is to ensure that the model can be trained.}

\item{features}{list of data.frames of all the features for the patients in the system going back in time. For example the first data.frame represents age 80, next is 75 and so forth. Each feature data.frame must contain an id column that matches the id column in the patient data.frame. The feature data.frame can contain any additional feature columns.}

\item{age_groups}{(optional) labels for the age groups. If the \code{patients} list is named, the age_groups will be the names of the lists.}
}
\value{
a MldpEHR object
}
\description{
Create MldpEHR object
}
\examples{
# Create a MldpEHR object
patients <- list(
    tibble::tibble(
        id = 1:10,
        age = 80,
        sex = sample(1:2, 10, replace = TRUE),
        followup = sample(0:5, 10, replace = TRUE),
        death = pmin(age + followup, sample(c(NA, 80:85), 10, replace = TRUE))
    ),
    tibble::tibble(
        id = 1:10,
        age = 75,
        sex = sample(1:2, 10, replace = TRUE),
        followup = sample(0:5, 10, replace = TRUE),
        death = pmin(age + followup, sample(c(NA, 75:80), 10, replace = TRUE))
    )
)

names(patients) <- c("80", "75")

features <- list(
    tibble::tibble(
        id = 1:10,
        feature1 = rnorm(10),
        feature2 = rnorm(10),
        feature3 = sample(1:3, 10, replace = TRUE)
    ),
    tibble::tibble(
        id = 1:10,
        feature1 = rnorm(10),
        feature2 = rnorm(10),
        feature4 = sample(0:1, 10, replace = TRUE)
    )
)

mldp <- MldpEHR(patients = patients, features = features)

}
