% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_markov_model.R
\name{.mortality_markov_model_for_stitch_model}
\alias{.mortality_markov_model_for_stitch_model}
\title{build an Markov probability model from multi-age prediction models}
\usage{
.mortality_markov_model_for_stitch_model(
  markov,
  model,
  step,
  qbins,
  required_conditions,
  min_obs_for_estimate = 10
)
}
\arguments{
\item{markov}{\itemize{
\item the markov model computed for the next age (older). the states in this age will be
mapped to the states in this markov layer
}}

\item{model}{\itemize{
\item prediction model (output of build_cross_validation_time_stitch_classification_models)
}}

\item{step}{\itemize{
\item time between prediction models
}}

\item{qbins}{\itemize{
\item quantile bin size of prediction score for which the markov model will define a state
}}

\item{required_conditions}{\itemize{
\item any filter to apply to the patients to filter out from model computation,
for example limiting the time window
}}

\item{min_obs_for_estimate}{\itemize{
\item minimum of observations required to compute probability per sex/sbin.
If minimum is not available, probability will be compuated using all data.
}}
}
\value{
a markov model, a list with the following members:
\itemize{
\item model - matrix containing the probability for each quantile(score) bin to reach each of
the target_classes provided in the oldest model.
\item local.model - data.frame containing the probability for each quantile(score) bin to reach
each of the quantile(score) bins of the next model by age.
\item qbins -  bins
\item target - data frame containing the target bin for this age model (to be used as outcome for
the younger age model)
}
}
\description{
build an Markov probability model from multi-age prediction models
}
