% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-prediction_model.R
\name{mldp_plot_multi_age_predictors_ecdf}
\alias{mldp_plot_multi_age_predictors_ecdf}
\title{Plot an ecdf of the score of each model for colored by outcome}
\usage{
mldp_plot_multi_age_predictors_ecdf(predictors)
}
\arguments{
\item{predictors}{a list of predictors, output of \code{mldp_mortality_multi_age_predictors}}
}
\value{
a ggplot object
}
\description{
For the oldest age the outcome is death, for younger ages the outcome is the score quantile of the next age group.
}
\examples{
# Load a small example data
data <- load_mortality_example_data(N = 100, num_age_groups = 3)
predictors <- mldp_mortality_multi_age_predictors(
    data@patients,
    data@features,
    survival_years = 5,
    nfolds = 2,
    q_thresh = 0.2,
    nthread = 2 # CRAN allows only 2 cores
)
mldp_plot_multi_age_predictors_ecdf(predictors)

}
