% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_model.R
\name{mldpEHR.prediction_model_features}
\alias{mldpEHR.prediction_model_features}
\title{analyze feature significance for a k-fold cross validation model using shaply values}
\usage{
mldpEHR.prediction_model_features(predictor)
}
\arguments{
\item{predictor}{\itemize{
\item classification model trained by mldpEHR.cv_train_outcome
}}
}
\value{
a list with the 3 objects
\itemize{
\item summary - data frame containing for each feature the mean absolute shaply value for the feature across all training data
\item shap_by_patient - data frame containing for each patient and feature the feature value and mean shap value across all training folds
\item shap_by_fold - similar to shap_by_patient, but for each fold seperately
}
}
\description{
analyze feature significance for a k-fold cross validation model using shaply values
}
\examples{
library(dplyr)
library(ggplot2)
target <- data.frame(id = 1:1000, sex = rep(c("male", "female"), 500), target_class = rep(c(0, 1), each = 500))
features <- data.frame(id = 1:500, a = rnorm(500), b = rnorm(500)) \%>\%
    bind_rows(
        data.frame(id = 501:1000, a = rnorm(500, mean = 0.5, sd = 2), b = rnorm(500, mean = -0.5, sd = 2))
    )
predictor <- mldpEHR.cv_train_outcome(target, features, folds = 3)
predictor_features <- mldpEHR.prediction_model_features(predictor)
}
